Program specific global variables for in options library
Thanks to global variables, constant variables can be used easily in any of
the functions of `lib/options.c'. So the program name, executable file and
bibtex record are now external variables in `lib/options.h' and global
variables in the program's `ui.c'.

With this addition, the `--onlyversion' and `--cite' options can be
immediately executed and allow the program to abort without going onto read
the rest of the options.

Also, to allow easy modification of a program's author list, a new global
variable is defined (in the renamed `authors-cite.h'). Later, we can change
this to `authors-cite.h.in' and during the making of the program, build it
to get the ordered list of contributing authors from the Git source, as we
do for the documentation.