Option descriptions also printed with values
Until now, only the option names and values were printed. But in many
cases, the option names alone are not descriptive enough and as a result,
you end up having to flip through the outputs of `--help' or
`--printparams'.

With this commit, the option's `doc' string (from `argp_options') is also
printed along with the name and value. Prior to actually printing the
values on `stdout' we first print the values in a string pointer, get all
the lengths, then find the maximum width to have all the comments
immediately under each other. As a result some modifications were made to
the functions to make this process easier for human readers and CPUs.