Other configuration files read immediately when seen
Until now, the other configuration files specified with the `--config'
option would all be packed together (only from the command-line) and read
before reading the configuration files. This had many limitations: it
wouldn't allow specifying another configuration file within existing
configuration files for example. Or, to read the configuration files
immediately on the command-line.

So, the `options.c' funcitons were slightly changed to allow the
configuration files specified with the `--config' option immediately when
they are read, similar to how we immediately check the version or print
citation.

To make things much more clear and managable, all global and external
variables were also removed and put in the `gal_options_common_params'
structure.

The documentation regarding options in the "Common program behavior" was
also reviewed and corrected to account for all the changes of the new
option management system.

This finishes task #14300.