All mandatory options not given are listed with error
Until now, as soon as a mandatory option without a value was found, an
error would be printed. So if multiple necessary options were not given,
then multiple runs were necessary to find all of them. But now, the names
and descriptions of such options are passed to linked lists and after
parsing all the options, they are all printed together, so in one run the
user can see all the necessary options. Also, the description of the option
is printed to help the user know what the problem was.

While doing this, the check for the FITS output name and `tabletype' option
in the Table program was moved only to when the `tabletype' parameter has
been set.

Also, it was noticed that in one of the rows, the spacing between the
string column of the numerical string was not enough. This is now
corrected.