Correction in alignment and getting pixel scale
The conditionals to get the WCS rotation matrix in ImageWarp's
`makealignmatrix' function were missing an `else' before the `altlin |= 2'
check! Thus when both flags where on, both conditions were executed which
would lead the to input matrix begin reset to zero. Therefore no alignment
would be done.

The reading of WCSLIB's `wcsprm' structure into a proper matrix is an issue
that is needed in other contexts too. Hence, the new function
`gal_wcs_array_from_wcsprm' was defined in `lib/wcs.c' to be a central
place for all Gnuastro to get this matrix when necessary and avoid such
bugs in the future.

This issue was submitted by Lee Kelvin as bug #50072 along with a FITS
image as an example. After this correction, the image would be rotated back
to its original aligned state. However, I noticed that the pixel scales had
changed! This would not happen for the FITS images that are produced by
running `make check'!

After going over each step, I noticed that the problem was in the
`gal_wcs_pixel_scale_deg' function, the pixel scale it returned was wrong
(even for the aligned image). Until now, the procedure to get the pixel
scale was very crude: we would use WCSLIB to get the coordinates of the
points (0,0), (0,1), and (1,0), then use `gal_wcs_angular_distance_deg' to
find the angular distance between the points. Since a very large number of
floating point calculations are necessary in the process, I guess it was
because of floating point errors that we got the wrong pixel scale for
Lee's image, but correct values for the others.

So, after reviewing my Linear Algebra and also some searches on the
internet, I found the robust solution to this problem: using `Singular
Value Decomposition'. Fortunately, the GNU Scientific Library had
implementations of the decomposition. Thus, as a side-effect to the bug
reported by Lee, `gal_wcs_pixel_scale_deg' now also returns the correct
pixel scale through a robust process.

This fixes bug #50072.