Decomposing PCi_j and CDELTi matices in output WCS
When reading images that have their WCS format in the `CDi_j' formalism,
WCSLIB will simply put the `CDi_j' array into the `PCi_j' array and set all
the `CDELTi' elements to `1.0'.

According to the FITS standard, in the `PCi_j' formalism, the scale factors
are encoded in the `CDELTi' keywords and the `PCi_j' matrix should contain
all the other transformations (recall that scaling is just one of all the
possible linear transformations).

Therefore, as noted by Lee Kelvin in bug #50073, this behavior of WCSLIB
can cause problems for programs that look into the value of `CDELTi' array
independent of the `PCi_j' matrix (like ds9 when viewing two files).

Fortunately, as part of the solution to bug #50072 (which was also reported
by Lee), a robust way to find the pixel scale was found using `Singular
Value Decomposition'. So when the `wcsprm' structure has `CDELTi' values of
`1.0', the new `gal_wcs_decompose_pc_cdelt' function is called to decompose
the two arrays so WCSLIB will print realistic `CDELTi' keywords.

These corrections also allowed for more clear pixel resolution checks in
ImageCrop's WCS mode crops.

This fixes bug #50073.