Always decompose PC and CDELT elements
Previously the decomposition of the `PCi_j' matrix and the `CDELTi' vector
would only be done when the elements of the latter were only 1. Either from
the input, or during the processing, the final values of `PCi_j' might
contain scale factors in them, independent of the `CDELTi' vector. For
example this is what ImageWarp currently does (it just modifies the `PCi_j'
matrix).

This was visible in the headers of `convolve_spatial_warped.fits' (after
running `make check'): The `PCi_j' diagonal matrix were a multiple of 5 and
the `CDELTi' was the original (input resolution).

So with this commit, the decomposition of `PCi_j' and `CDELTi' is done all
cases when a `PCi_j' matrix is present: the pixel scale is found, they are
unified, then the pixel scale in each dimension is put into `CDELTi' and
divided by `PCi_j'.