Specific suffixes for one modular warping in ImageWarp
Until now, ImageWarp would use the `_warped.fits' suffix for automatic
output. But its some time that it also allows modular warpings (like
rotation, alignment, scaling and so on). Thus, when the user only wants one
of these operations, it is much more meaningful if the suffix also reflects
that particular warp.

A check is now added, so when there is only one element in the modular
warpings linked list, the suffix will be determined based on its
type. Otherwise, it will fall back to the generic `--warped.fits' suffix.

A short description of the `--output' option and Automatic output was also
added to the first paragraph after the examples in the "Invoking ImageWarp"
section of the book.

This task was suggested by Lee Kelvin.

This finishes task #14309.