Using image naxes[n] when checkcenter is larger
Until now, the width of the region to check for a blank center in ImageCrop
was absolute (value to the `--checkcenter' option). As reported by Lee
Kelvin in bug #50099, this caused a crash when the cropped image size is
smaller than this width.

So with this commit, some checks are added so when the image width (in any
dimension, or value to `naxes[n]') is smaller than the check-center width,
the width (in that dimension) is used, not the value given by the user.

This fixes bug #50099.