ImageCrop's section syntax is inclusive in both bounds
Until now, ImageCrop's section syntax (for the `--section' option) was
inclusive for the first point, but not inclusive for the last point. This
was counter-intuitive and confusing for the user. So with this commit, the
last point is also inclusive. This was suggested in task #14313 by Lee
Kelvin.

Also, until now, users (not necessarily Gnuastro developers) could not post
bugs/tasks or comments on the respective tracker. This was very
inconvenient. Hence, the permissions on Savannah for these trackers were
changed so anyone who has a Savannah ID (not necessarily a Gnuastro
developer), can create an issue and/or submit a comment on one. The book
was also corrected to reflect this change. For anonymous users, the Support
tracker is still available.

This fixes task #14313.