tmpfs-config-make now has the programs that can be built
The data structure, new option management and table reading/writing has now
passed a defining stage and we will be correcting the separate programs to
use them. To be able to run `make check' its necessary to have all the
corrected programs configured, so until this branch is complete and merged
with master, as each program adopts the new systems, they will be enabled
at configure time in `tmpfs-config-make'.