Fixed automatic output checking in Arithmetic
When checking if the output name is set, a `==0' was missed! This has been
corrected with this commit. Also, the non-quiet outputs of Arithmetic have
been made more clear and the output filename is also now reported.

Finally in `tests/during-dev.sh' the old `onlydirconf' option name was
changed to the new `lastconfig' option.