ImageCrop's default suffix in past tense
As suggested by Lee Kelvin in task #14316, in most Gnuastro programs the
default automatic output suffix is in past tense (like `_rotated.fits', or
`_labeled.fits'). But in the case of ImageCrop, it was in the infinitive
tense (simply `_crop.fits').

The other programs were also checked and indeed only ImageCrop (the first
program in Gnuastro) was like this. The automatic output suffix for the
Arithmetic program (`_arith.fits') was also questionable, but that is a
little complicated, because the name "Arithmetic" isn't a verb. Probably we
should be thinking of a better name for this program. So for the time
being, it hasn't changed with this commit.

This finishes task #14316.