ImageCrop only checks center when crop is defined by its center
Until now, ImageCrop would check if the center of a crop is blank or not in
any case. This check makes sense when the crop is defined by its center
(for example from a catalog), but when the crop isn't defined by its center
(for example with the `--section', or `--polygon' options) this behavior
can be annoying and cause unexpected behavior.

Also, until now, you could only confirm if the crop's center was filled or
not (and thus the file was created or not) from the `0' or `1' value in the
third column of the log file or on stdout. The log file does have comments,
but on stdout, this was rather cryptic and not intuitively understood. As
suggested by Mosè in task #14318 this could also cause confusions.

Also, until now `--checkcenter' would only take values `>=3', so, to
disable checking we needed another option (`--keepblankcenter'). But now,
`--checkcenter=0' means that no center checking will be done. Therefore the
old `--keepblankcenter' option is no longer needed and was removed. Infact,
the default value of `--checkcenter' was set to `0' so by default no center
checking will be done. Only if the user explicitly gives it a larger value
on the command-line or in their configuration files will it be
activated. This is done based on the principle of least astonishment: for a
new user, not having an output is a strange behavior (even it is is
beneficial for them). After they learn about it, they can use it.

Finally, the appropriate sections of the book were reviewed to reflect this
new behavior. Most of these sections had not been changed from the time
they were created and were outdated (even before this commit). This has
been corrected now.

Other corrections that were done as part of this task are:

 - The `iscenterfilled' function of `crop.c' would go on trying to find
   blank values in the central region. However, CFITSIO already outputs the
   `anynul' variable for the same purpose. To be fair, until now we would
   check if ALL the central region was blank, not just that it had a single
   blank value. The reason this was changed is that the user can now ask
   for a one pixel central region if they want to.

 - In ImageCrop's `reportcrop' function, the message string is now allocated
   with `asprintf' to avoid any memory issues with static allocation.

 - A new code in ImageCrop's log file was created to show cases where the
   center wasn't checked.

 - ImageCrop did its own writing of WCS header keywords, so the recently
   added `gal_wcs_decompose_pc_cdelt' was implemented in it.

This finishes task #14318.