Option values directly set in library
Until now, the option values were allocated and stored in a void pointer in
the `value' field of `argp_option'. So when the option was not set by the
user or in configuration files, `value==NULL'. But this forced us to go
through the options one later time and put those values into the main
program structure, in this secondary loop, we would also have to check if
the options are mandatory or not. This is very frustrating and prone to
many bugs. So the new `mandatory' and `set' fields were created in the
`argp_option' structure and the `value' pointer now directly points to the
final location of the variable in the main program structure. These steps
effectively removed the need for any second loop after control comes out of
`gal_options_read_config_set', the program author can directly use the
values in their structure.

To do this, some major modifications were necessary, because we needed
access to the main program structure (`p'), and common parameters structure
(`cp') when the options array is defined. So the basic Argp global
variables like `argp_program_version', `argp_program_bug_address',
`args_doc', and `doc' were taken out of `args.h' and now directly defined
in `ui.c'. The `option_keys_enum' is also defined in `ui.c'. We now
"include" `args.h' and `commonopts.h' directly into the high-level
`ui_read_check_inputs_setup'.

Also to make the definition of the options more clear/readable, many new
enumerator types were defined and are now used in the programs.

Work on the MakeProfiles program is still ongoing and these changes were
all in line with diong that. So currently MakeProfiles just reaches the end
of `ui_read_check_inputs_setup'.