Cosmetic changes in option printing style
While finalizing the work on MakeProfiles, the style of printing option
printing style slightly changed to be more readable/useful. In particular,
when an option's `doc' string is longer than 80 characters, it is now
broken into multiple lines, similar to what `--help' does.

Because the `doc' strings can be broken into multiple lines, it is now
necessary to start the `doc' string with a `#', otherwise the first word of
the new line would be read as an option name, and the second as its value.

Also, thanks to the new features of Arithmetic (in particular `where'), the
old `inputascanvas' option was no longer necessary. If a user doesn't want
the pixel values, they can use it. The strategies are now described in the
book under the `--background' option.