CosmicCalculator working with changes in this branch
CosmicCalculator was one of the simplest applications, so its files (in
particular `args.h' and `ui.c' were updated to use the new option
management system. Also, a sanity check was added to check if the
fractional densities sum up to 1.0 or not. It now builds properly and its
simple test script in `make check' passes successfully.