Added -j8 to make examples in book
Compilation time can be greatly reduced if people use the `-j' option for
`make', which will distribute the `j'obs into multiple threads. So to help
the book readers, a `-j8' was added to the example scripts for all the
libraries and also in Quick start. A comment is added afterwards,
explaining that the `8' can be replaced with the number of CPU threads.

It could have been `-jN', but that would be a little cryptic and not work
if people didn't pay attention to the comment, making things a little
confusing. In this manner, even if they don't have 8 CPU threads, there
won't be any error and the speed will be up anyway (to the limit available
by their CPUs).