ConvertType working except for text inputs and outputs
ConvertType has been mostly re-written to effectively use `gal_data_t' and
all its added features in this branch. The only major visible changes
are:

 - The log `--log' option is removed (since it was the same as the common
   option `--log' option). In practice, people can use the Arithmetic to
   get the log of the image and then convert that. It would be interesting
   to implement it, but for now, I don't have much time.

 - Until now, the 8-bit images were inverted by default and this behavior
   was disabled with the `--noinvert' option. Having a `no...' option is
   very confusing both for the users and the developers. So it was changed
   to `--invert' and by default, there is no inversion.

Besides these, I also noticed that in some of the arithetic functions in
particular, we were mistakenly setting the output pointer based on the
input `void *' pointer, not based on the pointer to the proper type. This
caused problems because arithmetic on `void *' is like arithmetic on `char'
(1, not 8). So after an investigation, several such cases were corrected.

For all the tests written so far, ConvertType works nicely, we just need to
write library functions to read and write from text files to make the final
two tests pass.