Text library can write images also, besides tables
Until now, the text library was defined to only read/write plain text
tables (a collection of 1D data). But now, there are also functions to read
and write 2D (image) data to a plain text file. This was mainly done with
the motivation to complete ConvertType (so it can read from and write to
plain text files). But as a library function it is usable in many other
contexts also.

To be able to write 2D arrays as well as 1D columns in a text files, the
text table reading/writing functions were modularized (broken up) into
multiple functions and modified so the same infra-structure can be used.

In the process, some other corrections were also made that are described
below:

 - The FITS functions were renamed to follow a much more clear naming
   convention, for example the old `gal_fits_write_img' function is now
   `gal_fits_img_write'. Similarly, all functions related to FITS images
   start with `gal_fits_img', those starting with header keywords start
   with `gal_fits_key' and so on.

 - As part of the previous case, the functions related to the FITS WCS have
   been moved to the `gnuastro/wcs.h' header.

 - When no floating point printing format is specified by the user, the
   default format is changed to `%g' (for printf) and `E' (for FITS) to
   allow a more general demonstration of the data at a wider range.

 - The default width of printing single precision floating point values was
   updated to 13 (from 12) to allow for the default change above in
   printing the test table of `make check'.