Gnuastro now uses fixed width (from stdint.h) types
Until now the different types of `gal_data_t' were defined based on the
standard C types like `int', `short' and `long'. This was initially
modelled on CFITSIO's `TTYPE' macros like `TINT', or `TLONG'. But this
model is very hard to maintain for a portable system because on different
systems, `int' or `long' can have different widths. This would make the
code very hard to maintain. So following the types defined in `stdint.h',
the type constant names were changed and also, `stdint.h' types were used
to make sure we get a consistent result on all systems.

As a result of this clear definition of types based on widths, the
conversions from CFITSIO types (which are relative) to the `stdint.h' types
which are absolute is now much more clear.

In the process, a new `func' element was added to the `argp_option'
structure that will link to a function to parse the user given string and
put it in the given variable. `gal_options_read_type' is one function that
was defined for this job to read the string types given by the user and
convert them to internal integer codes. This was primarily done as part of
using the fruits of `gal_data_t' in ImageWarp, where multiple options (like
`--scale', `--rotate', and all other modular warpings) should be used to
update one linked list, but was then first tested on the `--type' option.