Option strings directly converted to internal codes
The `searchin' and `tableformat' options (like the `type' option), accept a
string from the user (on the command-line or in configuration files), but
then store the value internally as an integer for easy/fast checking and
further processing. Until now, the input string would be stored in the
`gal_options_common_params' and would go into the program for conversion to
the internal code.

Since `argp_option' now has a `func' field, it is now possible to define
one function that will immediately parse the user's given value and store
the string into `gal_options_common_params', so there is no more need to
keep the user's input string. The program authors can also be sure that
once option parsing is complete, the value is usable/recognized.