MakeNoise now uses the new gal_data_t infrastructure
The fruits of the new data structure are now implemented in MakeNoise. In
the process it was found that MakeNoise isn't reporting the correct seed
value to the random number generator, this problem has been fixed. It also
now reports the seed in verbose mode in any case (previously it would only
report the seed when an environment seed was to be used).

After finishing MakeNoise and running `make check' several of MakeProfiles
tests (that depended on MakeNoise outputs) were also tested and the
respective modifications were made in them to run successfully. In the
process, two corrections were made in MakeProfiles:

  - The old `--inputascanvas' is now was re-instated but with a new name:
    `--clearcanvas'. It was mainly removed because of its non-clean
    implementation before. But sice MakeProfiles' `ui.c' is now clean,
    re-implementing this feature was very easy. It can be very useful for
    practical purposes, when making labeled image for example.

  - Until now, MakeProfiles would abort with an error when the input
    columns had blank values. But for masking it might happen that you set
    a blank magnitude. So this check has now been removed when reading the
    magnitude column.