Statistics returns basic results, no array lib, new blank lib
The work on this commit started with merging the Statistics program with
the new data structure. However, it soon grew to the libraries. The full
process is described below.

The Statistics program can now read a table column as well as an image for
input. It can also remove certain ranges of data from its internal analysis
and as before also ignores blank pixels. Implementation of Statistics is
not yet complete and its section will be updated in the book after it is
complete (hopefully in the next commit). This commit was necessary because
too many changes were made in the libraries.

In order to do things clearly in statistics, many new operators were added
to the Arithmetic library (and thus the Arithmetic program): `numvalue',
`sumvalue', `meanvalue', `stdvalue', `medianvalue', `num', `mean' (instead
of the old `average'), and `std'.

Working with blank pixels until now was done in the main `data.h' library,
causing a bloating of operations in that library. To be clear and easy to
understand, the `data.h' is now only concerned with the most low-level
dataset issues (like types, allocation, copying, or reading and writing
single elements from strings). So we now have a new `blank.h' library for
functions related to blank pixels.

The old `array.h' library was a remnant of the past and with the new
libraries is now obsolete. So it was removed (its files and also its
sections in the book). Many of the old obsolete statistics functions have
also been removed.