Minor typo corrections in options of Header and Statistics
In header, the `--history' option was mistakenly given a `-h' short option
value! So you couldn't specify the extension (which comes from the common
options). It is now corrected to -H (as it was before).

In the Statistics program options, the appendix number in arXiv:1505.01664
which describes the mode finding algorithm was mistakenly written as
`D'. While it should have been `C'. This is also now corrected.