Image created on second HDU, common options to gnuastro.conf
In the FITS standard, the primary HDU can only be an image. So when it
makes tables, it puts them in the second HDU and makes a blank zero-sized
first HDU before the table HDU. Since we are working on making all the
programs work similarly on images and tables, this would easily confuse the
users and make the programs hard to run.

So now, when making a new FITS image, Gnuastro will also make a blank first
HDU and then put the image in the next HDU. This has several great
advantages besides the issue above: the user can use the blank header space
to set any extra keywords they like. In the FITS standard a HDU must
contain atleast 2880 characters or 36 keywords. There are 5 necessary
keywords that must go in that space and CFITSIO puts two comments, so in
total the user can put 36-7=29 keywords there without changing the file
size! With the new Fits program (old name: Header), the user can write any
extra information here. Another great side-effect is that when all the
files follow this convention, then it will feel as though we are counting
the HDUs from 1, not zero.

Also, as part of this commit, the new file `lib/gnuastro.conf' file has
been created which will also be installed as the system-wide common-option
configuration file. So all the similar options in the program configuration
files have been removed.

Another strange thing was that until now the `gnuastro.conf' file would be
read before each program's configuration file! This was contrary to the
increasing generality principle behind the configuration files. So for
example, if you wanted to use the `lastconfig' option for all the programs
in one directory, you would have to put it in each programs config file! So
now, it is reversed, and in each directory, first we read the program
configuration file and then `gnuastro.conf'. So only one `lastconfig' in
`gnuastro.conf' for example would be enough to stop all the programs from
parsing other configuration files.

Another small thing was that the `--quality' option of ConvertType wasn't
initialized! So when it wasn't called the error message wasn't useful!

Finally, `tests/prepconf.sh' and `tests/duringdev.sh' were also corrected
to accommodate this change.