Fits prints HDU info of FITS file with no options
Until now, the Fits program would print all the keywords in the given
extension when it was called with no option. But now that it is being
modified to also work on extensions, it now prints the extension
information with no option. To print all the keywords in a header, there is
the new `--printall' option.

The basic idea behind this change is simple: when the user doesn't give any
options (including specifying a HDU), then it is most probable that they
don't know the HDUs in the FITS file: how many there are, are they an image
or a table and so on. To get the full list of keywords, users can now call
the `--printall' option.

In the process, the source code for Fits was also re-organized to allow
this more general change of also manipulating extensions also. In the book
the old `Header' name has also been changed to `Fits' and the `--printall'
option has been added. The full behavior with the added feature to work on
extensions is not yet included in the manual. As soon as options to add and
remove HDUs are added, the book will also be updated.