HDU modification now fully implemented in Fits program
The Fits program now fully implements HDU modification options `--copy',
`--cut' and `--remove'. The Fits program section of the book has also been
completely re-written to accommodate these changes. Also, by default, the
Fits program will now print the extension information when no option is
specified. There is a new `--printallkeys' (`-p') option to print all the
keywords in one HDU. A test was also written to check the `--copy' option.

In the process, the `gal_checkset_check_remove_file' now accepts a `keep'
argument. This was done because unlike all the other programs until now,
the Fits program must not delete the output file if it is specified: we
want to add the HDU as an extension to the output file, not delete the
file.