Tesselation with the new data structure: work started
Work has started on implementing the tessellation processing with the new
`gal_data_t'. Thanks to its starter pointer and sizes, it was just
necessary to define a new `block' element. From now on, when `block' is
non-NULL, it means that the array in this dataset is not actually allocated
but is only one part of it.

The old `mesh' and `spatialconvolve' library source files have been removed
and we now have two new source files called `tile' and `convolve'. The old
routines in the former source files are being transferred to the new source
files.

An intial implementation of defining a full, non-overlapping tessellation
has been done in `gal_tile_all_position'. Using this function, we have
defined `gal_tile_all_position_two_layers' which will define the channels
and tiles tessellations. Previously this was done in a (very complicated)
mesh structure (`gal_mesh_params') that was not very trivial to navigate
and understand. The old structure was mainly focused on efficiency and only
defined for regular non-overlapping grids/tessellations, but was hard to
follow and less general.

The new system will allow arbitrary tiles defined anywhere (even
overlapping) over an allocated array. It is also extremely easy to
understand, since there is no new structure, it is the same old
`gal_data_t' that a developer is already fully familiar with and is
commonly used in many libraries/programs.

This was mainly done in preparation of porting NoiseChisel and SubtractSky
to `gal_data_t', but currently it is being tested on Convolve's
spatial-domain convolution. Once it has been fully tested and implemented
here, it will be really easy to apply it in those programs. Infact,
SubtractSky will probably be removed, because we can define tessellation
feature in Statistics which can effectively estimate the Sky value (and
many other statistics) on a grid. Then it is possible to use Arithmetic to
actually subtract the derived image from the input image.

The tile-related options (currently in Convolve, but later also in
NoiseChisel and Statistics) have also been updated to be more easier. Also,
in the tessellation of the full dataset, the remainder area would go into
the last tile in each dimension. But now it is the first tile that is
affected by any remainder area along each dimension and when the remainder
is significant, it will be divided between the first and last tiles.

Note that Convolve's spatial convolution is still not fully implemented yet
and its test will fail with this commit. The work done in Gnuastro has been
significant until this stage (defining the tessellation). So, this commit
is being made before starting work on each tile.

Some other modifications with this commit:

 - The functions that work with arrays of data structures have been
   separated as `gal_data_array_*', currently there is only two: one for
   clear-allocation and the other freeing the arrays and their contents.

 - Since the `--type' option is now a common option to all programs, a new
   `gal_fits_img_write_to_type' was defined to make the job easy in each
   program.

 - In many cases, we want all the flags for `gal_arithmetic', so a new
   `GAL_ARITHMETIC_FLAGS_ALL' was defined to make the codes more clear.

 - Since the programs are increasingly becoming dimension-independent, a
   `multidim' source library file has been defined to help in making the
   functions more clear to read.