Working on tiles implemented in tile checking
The most basic thing to do on tiles before actually working on them is to
check them. All the programs that use tiles have the `--checktiles'
(previously `--checkmesh') option for the user to see if their selected
tiles suite their purpose. So the `gal_tile_block_check_tiles' library
function was defined to do this and has been tested with this commit.

The function also surved as a great test-bed to define a general way to
parse a tile over an allocated block of memory. The old `lastmeshfrac'
option has also been renamed to `remainderfrac' to be more clear.

In the process the following changes were also made:

 - The two `gal_data_ptr_increment' and `gal_data_ptr_dist' have been
   defined to allow arithmetic on `void *' pointers given a certain
   type. These were necessary for working directy with the `array' element
   of `gal_data_t' (which is `void *').

 - The `gal_multidim_add_coords' function was added for easy addition of
   coordinates (with any dimension).