Spatial convolution now using new tessellation
Spatial convolution (on multiple threads) is one of the most basic and
common applications of the tessellation management that was designed in the
previous commits. With this commit convolution on central tiles of the
image has been applied. Only the edge tiles remain. The new system has
significantly simplified the code (and probably the efficiency) of spatial
convolution in Gnuastro.

Also, reading of the input image has now been changed to float32
type. Previously it was float64, while the kernel was initially read as
float64. For frequency domain convolution this makes no difference, because
the input is put into a complex float64 type array before the processing is
done. For spatial domain convolution it also makes no difference because
the sum is stored as float64.

Finally, as before with the `mesh' library, a new
`gal_tile_function_on_threads' function has been defined to make it easy to
spin off threads working on separate tiles.