The noedgecorrection option in spatial domain convolution
In the old implementation of spatial convolution we had a
`--noedgecorrection' option to avoid corrections on the edge of a
region. But it was not implemented in the new re-write. So with this
commit, it is included. The description of the old `mesh' options were also
removed from the `Mesh grid section' of the book (in the previous commit we
the new tile options are discussed as part of the common options).

Within the Convolve program, edge correction is the default in the spatial
domain and the user can disable it with the `--noedgecorrection'
option. However, in the library function `gal_convolve_spatial', the user
must pass a value of `1' to the `edgecorrection' argument to have edge
correction. This is done because a library user is generally looking for a
more raw/generic processing, while the program user will usually choose
spatial domain over frequency domain because of the edge correction
feature, so this is more convenient for them. Ofcourse the program behavior
can easily be changed to the library behavior if there is demand.

In the process, I tried compiling without the debugging flag and
optimization. The compiler made some interesting new warnings about
non-initialized variables that are also now corrected.