Statistical functions can accept tiled input
The main statistical functions (currenly on the single measurement class in
the Statistics program) are now equipt to deal with inputs as tiles. This
will be correced in the next commit. This was made possible thanks to the
new `GAL_TILE_PARSE_OPERATE' macro that will parse a tile and do a given
operation. However, when channels are defined, the array keeping a single
value for a tile will not be ordered properly.

In the process, the following modifications were also made:

 - Statistics now takes a `mirrordist' option for mode-related
   calculations. Until now, this was hardcoded.

 - `gal_blank_present' can now work with tiles also, infact this was the
   first function to be modified for this job.

 - `gal_tile_to_contiguous' will copy the contents of a tile into a
   contiguous patch of memory.

 - `gal_statistics_no_blank_sorted' now also puts the tile's values into a
   contiguous piece of memory before removing blanks and sorting! In
   principle, these two operations cannot be defined in a non-contiguous
   patch of memory.