gal_data_copy_* functions can now work on tiles
In order to continue work on the tiles and their interpolation some small
separate corrections were made. The most important is that the
`gal_tile_to_contiguous' function that was introduced in the previous
commit has now been removed. The `gal_data_copy_to_new_type' function (and
thus all the other `gal_data_copy...' functions can now work on tiles (copy
a tile to a contiguous patch of memory of any given type), so there was no
more need for `gal_tile_to_contiguous'.

The other changes are:

 - The name `gal_wcs_array_from_wcsprm' was ambiguous so it has been
   renamed to `gal_wcs_warp_matrix'. This was done as part of work to
   prepare the WCS properties of a tile.

 - `gal_data_copy_wcs' (in `gnuastro/data.h') has been replaced by
   `gal_wcs_copy' in the `gnuastro/wcs.h' header.

 - `gal_data_to_same_type' has been completely removed since it was very
   specific to some special condition that doesn't exist any more (this
   function isn't used any part of Gnuastro). The later defined
   `gal_data_copy_to_new_type_free' (in combination with
   `gal_data_copy_to_new_type') can be easily used instead.

 - `gal_data_out_type' was just copy and pasted to a more relevant part of
   `lib/data.c' (to do with type information). Until now it was near the
   copy functions.

 - `gal_wcs_on_tile' will now copy the WCS of the tile's allocated block
   and correct it for use in the tile.

 - The default settings for the starting and ending (inclusive) elements of
   a tile were corrected (subtracted by one since it is meant to be
   inclusive: to avoid bugs when changing dimensions which can lead to
   large increments).