Structure to keep tessellation information
The Convolve program was the first to benefit from the new tessellation
infra-structure of `lib/tile.c'. However, it was arguably the easiest
application: it did not need to manage a single value for each tile and
interpolation for example. However, now that the tile structure is being
implemented in Statistics, we have a single value for each tile. Given the
possiblity of multiple channels, the tile IDs and 2/3D positions don't
simply correspond to each other and many pieces of information about the
tessellation are necessary to deduce the proper conversion.

So to make things easier, a `gal_tile_two_layer_params' structure was
defined in `lib/gnuastro/tile.h' keeping all the input and internally
necessary parameters to avoid having copy around different arrays. Many
things got much more cleaner now and easier to manage. Allowing us easily
to convert tile IDs into positions in the dataset and vice-versa which is
needed for the final interpolation.

`gal_tile_two_layer_params' is now part of `gal_options_common_params' and
thus available to any program. Thus the main program parameter
initialization has been corrected in all the programs to also initialize
this structure's values. Library users can also easily fill it up and call
`gal_tile_full_two_layers' to generate the two layer tessellation. This
will be explained in the manual.

Finally, having this structure also allowed the creation of
`gal_tile_full_free_contents' to easily free all the allocated space for
the tessellation.